requirements.txt
autoenv==1.0.0
click==7.1.2
DateTime==4.3
Flask==1.1.2
Flask-Cors==3.0.10
flask-mysql-connector==1.0.3
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.23
mysqlclient==2.0.3
protobuf==3.14.0
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
waitress==1.4.4
Werkzeug==1.0.1
zope.interface==5.2.0
